+++ lib/components/ItemContent.tsx
